Fix Button styles, ability to pass function into the "text" prop
